+++ tests/test_file_hashing.py
@@ -1,7 +1,7 @@
 import base64
 from pathlib import Path
 
-from lamindb._dobject import hash_file, to_b64_str
+from lamindb._record import hash_file, to_b64_str
 from lamindb.dev._core import get_name_suffix_from_filepath
